Refactor : List Header - remove classname split Components
